Instanciated context

+++ ru.kmorozov.gbd.core/src/main/java/ru/kmorozov/gbd/core/logic/extractors/PageImgProcessor.java
@@ -4,6 +4,7 @@ import ru.kmorozov.gbd.core.config.GBDOptions;
 import ru.kmorozov.gbd.core.logic.Proxy.AbstractProxyListProvider;
 import ru.kmorozov.gbd.core.logic.Proxy.HttpHostExt;
 import ru.kmorozov.gbd.core.logic.connectors.Response;
+import ru.kmorozov.gbd.core.logic.context.BookContext;
 import ru.kmorozov.gbd.core.logic.model.book.PageInfo;
 import ru.kmorozov.gbd.core.utils.Images;
 import ru.kmorozov.gbd.core.utils.Logger;
@@ -11,14 +12,12 @@ import ru.kmorozov.gbd.core.utils.Logger;
 import java.io.*;
 import java.net.SocketTimeoutException;
 
-import static ru.kmorozov.gbd.core.logic.ExecutionContext.INSTANCE;
+import static ru.kmorozov.gbd.core.logic.context.ExecutionContext.INSTANCE;
 
 /**
  * Created by km on 21.11.2015.
  */
 class PageImgProcessor extends AbstractHttpProcessor implements Runnable {
-
-    private static final Logger logger = Logger.getLogger(INSTANCE.getOutput(), PageImgProcessor.class.getName());
 
     private static final String IMG_ERROR_TEMPLATE = "No img at %s with proxy %s";
 
@@ -26,23 +25,25 @@ class PageImgProcessor extends AbstractHttpProcessor implements Runnable {
 
     private final PageInfo page;
     private final HttpHostExt usedProxy;
+    private final BookContext bookContext;
+    private final Logger logger;
 
-    public PageImgProcessor(PageInfo page, HttpHostExt usedProxy) {
+    public PageImgProcessor(BookContext bookContext, PageInfo page, HttpHostExt usedProxy) {
+        this.bookContext = bookContext;
         this.page = page;
         this.usedProxy = usedProxy;
+        this.logger = INSTANCE.getLogger(PageImgProcessor.class, bookContext);
     }
 
     private boolean processImage(String imgUrl, HttpHostExt proxy) {
-        if (GBDOptions.secureMode() && proxy.isLocal())
-            return false;
+        if (GBDOptions.secureMode() && proxy.isLocal()) return false;
 
         File outputFile = null;
         Response resp = null;
         InputStream inputStream = null;
         OutputStream outputStream = null;
 
-        if (page.loadingStarted.get())
-            return false;
+        if (page.loadingStarted.get()) return false;
 
         try {
             resp = getContent(imgUrl, proxy, false);
@@ -62,29 +63,21 @@ class PageImgProcessor extends AbstractHttpProcessor implements Runnable {
                     String imgFormat = Images.getImageFormat(resp);
                     if (imgFormat != null) {
 
-                        if (page.loadingStarted.get())
-                            return false;
+                        if (page.loadingStarted.get()) return false;
 
                         page.loadingStarted.set(true);
-                        outputFile = new File(INSTANCE.getOutputDir().getPath() + "\\" + page.getOrder() + "_" + page.getPid() + "." + imgFormat);
+                        outputFile = new File(bookContext.getOutputDir().getPath() + "\\" + page.getOrder() + "_" + page.getPid() + "." + imgFormat);
 
-                        if (reloadFlag = outputFile.exists())
-                            if (GBDOptions.reloadImages())
-                                outputFile.delete();
-                            else
-                                return false;
-                    } else
-                        break;
+                        if (reloadFlag = outputFile.exists()) if (GBDOptions.reloadImages()) outputFile.delete();
+                        else return false;
+                    } else break;
 
-                    if (outputFile != null && outputFile.exists())
-                        break;
+                    if (outputFile != null && outputFile.exists()) break;
 
                     if (!proxy.isLocal())
-                        logger.info(String.format("Started img %s for %s with %s Proxy",
-                                reloadFlag ? "RELOADING" : "processing", page.getPid(), proxy.toString()));
+                        logger.info(String.format("Started img %s for %s with %s Proxy", reloadFlag ? "RELOADING" : "processing", page.getPid(), proxy.toString()));
                     else
-                        logger.info(String.format("Started img %s for %s without Proxy",
-                                reloadFlag ? "RELOADING" : "processing", page.getPid()));
+                        logger.info(String.format("Started img %s for %s without Proxy", reloadFlag ? "RELOADING" : "processing", page.getPid()));
 
                     outputStream = new FileOutputStream(outputFile);
                 }
@@ -105,8 +98,7 @@ class PageImgProcessor extends AbstractHttpProcessor implements Runnable {
             return true;
         } catch (SocketTimeoutException ste) {
             proxy.registerFailure();
-        }
-        catch (Exception ex) {
+        } catch (Exception ex) {
             ex.printStackTrace();
         } finally {
             if (inputStream != null) {
@@ -130,8 +122,7 @@ class PageImgProcessor extends AbstractHttpProcessor implements Runnable {
             }
 
             try {
-                if (resp != null)
-                    resp.close();
+                if (resp != null) resp.close();
             } catch (IOException e) {
                 e.printStackTrace();
             }
@@ -141,24 +132,21 @@ class PageImgProcessor extends AbstractHttpProcessor implements Runnable {
     }
 
     private boolean processImageWithProxy(HttpHostExt proxy) {
-        return !(!proxy.isLocal() && !proxy.isAvailable()) && processImage(page.getImqRqUrl(ImageExtractor.HTTPS_TEMPLATE, GBDOptions.getImageWidth()), proxy);
+        return !(!proxy.isLocal() && !proxy.isAvailable()) && processImage(page.getImqRqUrl(bookContext.getBookInfo().getBookId(), ImageExtractor.HTTPS_TEMPLATE, GBDOptions.getImageWidth()), proxy);
     }
 
     @Override
     public void run() {
-        if (page.dataProcessed.get())
-            return;
+        if (page.dataProcessed.get()) return;
 
         if (!processImageWithProxy(usedProxy)) {
             // Пробуем скачать страницу с без прокси, если не получилось с той прокси, с помощью которой узнали sig
-            if (!usedProxy.isLocal())
-                if (!processImageWithProxy(HttpHostExt.NO_PROXY))
-                    // Пробуем скачать страницу с другими прокси, если не получилось с той, с помощью которой узнали sig
-                    AbstractProxyListProvider.getInstance().getParallelProxyStream().forEach(proxy -> {
-                        if (proxy != usedProxy)
-                            if (processImageWithProxy(proxy)) {
-                            }
-                    });
+            if (!usedProxy.isLocal()) if (!processImageWithProxy(HttpHostExt.NO_PROXY))
+                // Пробуем скачать страницу с другими прокси, если не получилось с той, с помощью которой узнали sig
+                AbstractProxyListProvider.getInstance().getParallelProxyStream().forEach(proxy -> {
+                    if (proxy != usedProxy) if (processImageWithProxy(proxy)) {
+                    }
+                });
         }
     }
 }
